Extract xref utils into own file.
