+++ src/uniparc/components/entry/UniParcFeaturesView.tsx
@@ -9,7 +9,7 @@ import useDatabaseInfoMaps from '../../../shared/hooks/useDatabaseInfoMaps';
 
 import externalUrls from '../../../shared/config/externalUrls';
 import { stringToColour } from '../../../shared/utils/color';
-import { processUrlTemplate } from '../../../uniprotkb/components/protein-data-views/XRefView';
+import { processUrlTemplate } from '../../../shared/utils/xrefs';
 import { sortByLocation } from '../../../uniprotkb/utils';
 
 import { SequenceFeature } from '../../adapters/uniParcConverter';
